Extract ChainSync client state to own file
